requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
tqdm
requests
python-dotenv
scipy